+++ packages/core/solidity/src/generate/account.ts
@@ -1,0 +1,20 @@
+import type { AccountOptions } from '../account';
+import { accessOptions } from '../set-access-control';
+import { infoOptions } from '../set-info';
+import { upgradeableOptions } from '../set-upgradeable';
+import { generateAlternatives } from './alternatives';
+
+const account = {
+  name: ['MyAccount'],
+  accountBase: ['Account', 'AccountCore'] as const,
+  signer: ['ERC7702', 'ECDSA', 'P256', 'RSA'] as const,
+  ERC7821: [false, true] as const,
+  ERC7579: [false, 'AccountERC7579', 'AccountERC7579Hooked'] as const,
+  access: accessOptions,
+  upgradeable: upgradeableOptions,
+  info: infoOptions,
+};
+
+export function* generateAccountOptions(): Generator<Required<AccountOptions>> {
+  yield* generateAlternatives(account);
+}